requirements.txt
Django
djangorestframework
celery[redis]
redis
yfinance
torch
uvicorn[standard]
psycopg[binary]


